Added getters to get all categories and get all products within a category.

+++ app/src/main/java/com/example/ofn/data/model/Product.kt
@@ -4,6 +4,6 @@ data class Product (
     val productName: String,
     val category: String,
     val description: String,
-    val stock: Int,
+    val stock: Long,
     val imageUrl: String
 )
